// convocatoria-frontend/src/app/core/interfaces/evento.ts
import { Time } from '@angular/common';
import { Auditoria } from './auditoria';
import { Caracter } from './caracter';
import { Entidad } from './entidad';
import { Lugar } from './lugar';
import { Persona } from './persona';
import { TipoActoCulto } from './tipoActoCulto';
import { Titular } from './titular';

export interface Evento extends Auditoria {
  id: number;
  eventoPadre?: Evento;
  tipoActoCulto: TipoActoCulto;
  tipo?: string;
  edicion?: number;
  aniversario?: number;
  dia?: number;
  entidadOrganizadora: Entidad;
  caracter?: Caracter;
  fecha: string;
  horaIni1: Time;
  horaFin1?: Time;
  horaIni2?: Time;
  horaFin2?: Time;
  lugar: Lugar;
  titulares?: Titular[];
  personas?: Persona[];
  bandas?: Entidad[];
  marco?: string;
  titulo?: string;
  tituloEdicion?: number;
  hito?: string;
  hitoAniversario?: number;
  informacion?: string;
  itinerario?: string;
  nota?: string;
  visible: boolean;
  aplazado: boolean;
  suspendido: boolean;
  extraordinario: boolean;
  pendienteRevisar: boolean;
  revisarNota?: string;
  duplicado?: string;
  regla?: string;
  periodico: boolean;
  revisarTipoActoCulto: boolean;
  revisarDias: boolean;
  revisarLugar: boolean;
  revisarTitulares: boolean;
  revisarPersonas: boolean;
  revisarBandas: boolean;
  revisarInformacion: boolean;
  revisarItinerario: boolean;
  revisarRegla: boolean;
  duracion?: number;
  estado?: string;
}